Cleaned up file structure, added about link to NavbarAuthed.js, and info link to Footer.js
