+++ client/src/components/Shared/NavbarAuthed.js
@@ -96,6 +96,10 @@ class NavbarB extends Component {
               Product
             </NavLink>
 
+            <NavLink exact to="/about" activeClassName="activeA">
+              About
+              </NavLink>
+
             <NavLink exact to="/pricing" activeClassName="activeA">
               Pricing
             </NavLink>
