+++ client/src/views/App/About.js
@@ -1,14 +1,11 @@
 import React, { Component } from "react";
 import NavBarB from "../../components/Shared/NavbarB.js";
 import Search from "../../components/LandingSearch/Search.js";
-// import Footer from '../../components/HomeView/Footer';
-import "../../scss/Landing.scss";
 import Map from "../../components/HomeView/Map/Map.js";
-// import Suggestions from '../../components/LandingSearch/Suggestions.js';
-// const { API_KEY } = process.env
-// const API_URL = 'http://api.musicgraph.com/api/v2/artist/suggest'
+import Component1 from "../../components/AboutView/Component1.js";
+import Component2 from "../../components/AboutView/Component2.js";
 
-class Landing extends Component {
+class About extends Component {
   render() {
     return (
       <div>
@@ -16,10 +13,19 @@ class Landing extends Component {
         <div style={{ display: "none" }}>
           <Map />
         </div>
+        <div className="container">
+          <Component1 />
+          <Component2 />
+          <div className="header" />
+          <div id="main">
+            <div className="inner">
+              <div className="section" />
+            </div>
+          </div>
+        </div>
       </div>
-      // </div>
     );
   }
 }
 
-export default Landing;
+export default About;
